Create separate file for Game and Author's methods
