+++ main.rb
@@ -9,8 +9,8 @@ require './classes/author'
 require './classes/game'
 require './managers/book_manager'
 require './managers/music_manager'
+require './managers/game_manager'
 
-items = []
 def save_items_to_json(items)
   serialized_items = items.map(&:to_json)
   write_json_to_file('./data/items.json', serialized_items)
@@ -42,12 +42,6 @@ end
 
 items = load_items_from_json
 
-def list_genres(items)
-  puts 'List of all genres:'
-  genres = list_items(Genre, items)
-  genres.each { |genre| puts "Genre ID: #{genre.id}, Name: #{genre.name}" }
-end
-
 def list_items(class_name, items)
   items.select { |item| item.is_a?(class_name) }
 end
@@ -55,47 +49,6 @@ end
 def add_item(item, items)
   items << item
   puts "#{item.class.name} added successfully."
-end
-
-def list_authors(items)
-  puts 'List of all authors:'
-  authors = items.select { |item| item.is_a?(Game) }.map(&:author)
-  puts 'Author list is empty, please add a game first.' if authors.empty?
-  authors.each do |author|
-    puts "Author ID: #{author.id}, First Name: #{author.first_name}, Last Name: #{author.last_name}"
-  end
-end
-
-def list_games(items)
-  puts 'List of all games:'
-  games = list_items(Game, items)
-  puts 'List of games is empty, please add a new game.' if games.empty?
-  games.each_with_index { |game, index| puts "Game ID: #{index + 1}, Title: #{game.title}" }
-end
-
-def add_game(items)
-  puts 'Enter Game Title:'
-  title = gets.chomp
-
-  puts 'Is it Multiplayer? (true/false):'
-  multiplayer = gets.chomp == 'true'
-
-  puts 'Enter the Publish Date (YYYY-MM-DD):'
-  publish_date = Date.parse(gets.chomp)
-
-  puts 'Enter the Last Played Date (YYYY-MM-DD):'
-  last_played_at = Date.parse(gets.chomp)
-
-  puts 'Enter the Author First Name:'
-  first_name = gets.chomp
-
-  puts 'Enter the Author Last Name:'
-  last_name = gets.chomp
-
-  author = Author.new(items.size + 1, first_name, last_name)
-  new_game = Game.new(title, publish_date, multiplayer, last_played_at, author)
-
-  add_item(new_game, items)
 end
 
 puts 'Welcome to the Cataloge of my things'
