+++ managers/game_manager.rb
@@ -1,0 +1,41 @@
+def add_game(items)
+  puts 'Enter Game Title:'
+  title = gets.chomp
+
+  puts 'Is it Multiplayer? (true/false):'
+  multiplayer = gets.chomp == 'true'
+
+  puts 'Enter the Publish Date (YYYY-MM-DD):'
+  publish_date = Date.parse(gets.chomp)
+
+  puts 'Enter the Last Played Date (YYYY-MM-DD):'
+  last_played_at = Date.parse(gets.chomp)
+
+  puts 'Enter the Author First Name:'
+  first_name = gets.chomp
+
+  puts 'Enter the Author Last Name:'
+  last_name = gets.chomp
+
+  author = Author.new(items.size + 1, first_name, last_name)
+  add_item(author, items)
+
+  new_game = Game.new(title, publish_date, multiplayer, last_played_at, author)
+  add_item(new_game, items)
+end
+
+def list_games(items)
+  puts 'List of all games:'
+  games = list_items(Game, items)
+  puts 'List of games is empty, please add a new game.' if games.empty?
+  games.each_with_index { |game, index| puts "Game ID: #{index + 1}, Title: #{game.title}" }
+end
+
+def list_authors(items)
+  puts 'List of all authors:'
+  authors = list_items(Author, items)
+  puts 'Author list is empty, please add a game first.' if authors.empty?
+  authors.each do |author|
+    puts "Author ID: #{author.id}, First Name: #{author.first_name}, Last Name: #{author.last_name}"
+  end
+end
